+++ frontend/controller/actions/group.js
@@ -23,13 +23,11 @@ import { VOTE_FOR } from '@model/contracts/shared/voting/rules.js'
 import sbp from '@sbp/sbp'
 import {
   LOGOUT,
-  KV_QUEUE,
   OPEN_MODAL,
   REPLACE_MODAL,
   SWITCH_GROUP,
   JOINED_GROUP
 } from '@utils/events.js'
-import { KV_KEYS } from '@utils/constants.js'
 import { imageUpload } from '@utils/image.js'
 import { GIMessage } from '~/shared/domains/chelonia/chelonia.js'
 import { findKeyIdByName } from '~/shared/domains/chelonia/utils.js'
@@ -460,7 +458,7 @@ export default (sbp('sbp/selectors/register', {
               }
             })
 
-            await sbp('gi.actions/group/updateLastLoggedIn', {
+            await sbp('gi.actions/kv/updateLastLoggedIn', {
               contractID: params.contractID
             }).catch((e) => console.error('[gi.actions/group/join] Error sending updateLastLoggedIn', e))
           } catch (e) {
@@ -952,36 +950,6 @@ export default (sbp('sbp/selectors/register', {
 
     return sendMessage(params)
   }),
-  'gi.actions/group/updateLastLoggedIn': async ({ contractID }: { contractID: string }) => {
-    try {
-      const rootState = sbp('state/vuex/state')
-      const userID = rootState.loggedIn?.identityContractID
-
-      if (!userID) {
-        throw new Error('Unable to update last logged in without an active session')
-      }
-      const now = new Date().toISOString()
-
-      // Wait for any pending operations (e.g., sync) to finish
-      await sbp('chelonia/queueInvocation', contractID, () => {})
-
-      sbp('okTurtles.eventQueue/queueEvent', KV_QUEUE, async () => {
-        const fnGetUpdatedLastLoggedIn = async (cID, key) => {
-          const current = (await sbp('chelonia/kv/get', cID, key))?.data || {}
-          return { ...current, [userID]: now }
-        }
-
-        const data = await fnGetUpdatedLastLoggedIn(contractID, KV_KEYS.LAST_LOGGED_IN)
-        await sbp('chelonia/kv/set', contractID, KV_KEYS.LAST_LOGGED_IN, data, {
-          encryptionKeyId: sbp('chelonia/contract/currentKeyIdByName', contractID, 'cek'),
-          signingKeyId: sbp('chelonia/contract/currentKeyIdByName', contractID, 'csk'),
-          onconflict: fnGetUpdatedLastLoggedIn
-        })
-      })
-    } catch (e) {
-      throw new GIErrorUIRuntimeError(L('Failed to update "lastLoggedIn" in a group profile.'), { cause: e })
-    }
-  },
   ...encryptedAction('gi.actions/group/payment', L('Failed to create payment.')),
   ...encryptedAction('gi.actions/group/paymentUpdate', L('Failed to update payment.')),
   ...encryptedAction('gi.actions/group/sendPaymentThankYou', L('Failed to send a payment thank you note.')),
